Problem: unfair distribution of articles
Solution: it turns out that generating to indices favors articles with a
fewer quotes, which will get selected more often than others. In order
to avoid that the randomness is computed for all the quotes and the
Articles struct is traversed one by one in order to find a proper quote.

This way each quote gets its fair share for a display.